% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_inflation.R
\name{adj_inflation}
\alias{adj_inflation}
\title{A Function to get inflation-adjusted wages.}
\usage{
adj_inflation(base_year)
}
\arguments{
\item{base_year}{The base year of your choice (4 digits-year without quotation mark).}
}
\value{
wages and wages_hs_do data with a new column of inflation-adjusted wages value.
}
\description{
This function aims to adjust the \code{wage} column in \code{wages} and  \code{wages_hs_do} with the
inflation rate of the base year of your choice.
The inflation rate is calculated from the CPI data published by the US Bureau of Labor Statistics (BLS).
This function needs internet connection to download the data from the BLS website.
The code is adapted from Kris Eberwein in R Bloggers website. https://www.r-bloggers.com/2015/12/calculate-inflation-with-r/
}
\examples{
\dontrun{
   inflation_adjust(1990)
}

}
