% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog_nlsy79.R
\docType{data}
\name{demog_nlsy79}
\alias{demog_nlsy79}
\title{Demographic Data from the National Longitudinal Survey of Youth (NLSY79)}
\format{
A data frame contains 12,686 rows and 8 variables:
\describe{
\item{id}{A unique individual's ID number.}
\item{age_1979}{The age of the subject in 1979.}
\item{gender}{Gender of the subject: F = Female and M = Male.}
\item{race}{Race of the subject: NB,NH = NON-BLACK,NON-HISPANIC; H = HISPANIC; B = BLACK.}
\item{hgc}{Highest grade completed.}
\item{hgc_i}{Integer of highest grade completed.}
\item{hgc_1979}{The highest grade completed in 1979 (integer value).}
\item{ged}{Wether the respondent had a high school diploma or Graduate Equivalency Degree (GED). 1: High school diploma; 2: GED; 3: Both}
}
}
\source{
The U.S. Bureau of Labor Statistics. (2021, January 6). \emph{National Longitudinal Survey of Youth 1979}. https://www.nlsinfo.org/content/cohorts/nlsy79/get-data
}
\description{
A data set contains demographic information of the NLSY79 cohort.
The cohort included 12,686 respondents ages 14-22 when first interviewed in 1979.
}
