% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yowie-package.R
\docType{package}
\name{yowie-package}
\alias{yowie}
\alias{yowie-package}
\title{yowie: Longitudinal Wages Data from The NLSY79}
\description{
This package provides longitudinal wages data sets along with several demographic variables from The National Longitudinal Survey of Youth (NLSY) from survey year 1979 to 2018. There are three data sets in this package: 1. The wages data from the cohort whose highest grade completed is up to 12th grade; 2. The high school dropouts, and; 3. The demographic data of the NLSY79 cohort.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/numbats/yowie}
  \item Report bugs at \url{https://github.com/numbats/yowie/issues}
}

}
\author{
\strong{Maintainer}: Dewi Amaliah \email{dlamaleeah@gmail.com}

Authors:
\itemize{
  \item Di Cook \email{dicook@monash.edu}
}

}
\keyword{internal}
