% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yowie-package.R
\docType{package}
\name{yowie-package}
\alias{yowie}
\alias{yowie-package}
\title{yowie: Longitudinal Wages Data from the National Longitudinal Survey of Youth 1979}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Longitudinal wages datasets and several demographic variables from the National Longitudinal Survey of Youth from 1979 to 2018. There are three data sets in this package: The wages data from the NLSY79 cohort; The wages data of the high school dropouts and; The demographic data of NLSY79 cohort.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/numbats/yowie}
  \item Report bugs at \url{https://github.com/numbats/yowie/issues}
}

}
\author{
\strong{Maintainer}: Dewi Amaliah \email{dlamaleeah@gmail.com}

Authors:
\itemize{
  \item Di Cook \email{dicook@monash.edu}
  \item Emi Tanaka \email{emi.tanaka@monash.edu}
}

Other contributors:
\itemize{
  \item Kate Hyde \email{hyde.kate.a@gmail.com} [contributor]
  \item Nicholas Tierney \email{nicholas.tierney@gmail.com} [contributor]
}

}
\keyword{internal}
